requirements.txt
sanic == 25.3
python-dotenv
pymongo
PyJWT
passlib
jsonschema